用户信息迁移到setting
