+++ src/renderer/src/store/runtime.ts
@@ -26,17 +26,10 @@ export interface RuntimeState {
   update: UpdateState
   webdavSync: WebDAVSyncState
   export: ExportState
-  user: UserState
 }
 
 export interface ExportState {
   isExporting: boolean
-}
-
-export interface UserState {
-  username: string
-  token: string | null
-  isLoggedIn: boolean
 }
 
 const initialState: RuntimeState = {
@@ -60,11 +53,6 @@ const initialState: RuntimeState = {
   },
   export: {
     isExporting: false
-  },
-  user: {
-    isLoggedIn: false,
-    username: '',
-    token: null
   }
 }
 
@@ -98,9 +86,6 @@ const runtimeSlice = createSlice({
     },
     setExportState: (state, action: PayloadAction<Partial<ExportState>>) => {
       state.export = { ...state.export, ...action.payload }
-    },
-    setUserState: (state, action: PayloadAction<Partial<UserState>>) => {
-      state.user = { ...state.user, ...action.payload }
     }
   }
 })
@@ -114,8 +99,7 @@ export const {
   setResourcesPath,
   setUpdateState,
   setWebDAVSyncState,
-  setExportState,
-  setUserState
+  setExportState
 } = runtimeSlice.actions
 
 export default runtimeSlice.reducer
