+++ src/renderer/src/store/settings.ts
@@ -6,6 +6,12 @@ export type SendMessageShortcut = 'Enter' | 'Shift+Enter' | 'Ctrl+Enter' | 'Comm
 
 export type SidebarIcon = 'assistants' | 'agents' | 'paintings' | 'translate' | 'minapp' | 'knowledge' | 'files'
 
+export interface UserState {
+  username: string
+  token: string | null
+  isLoggedIn: boolean
+}
+
 export const DEFAULT_SIDEBAR_ICONS: SidebarIcon[] = [
   'assistants',
   'agents',
@@ -73,6 +79,7 @@ export interface SettingsState {
   thoughtAutoCollapse: boolean
   notionAutoSplit: boolean
   notionSplitSize: number
+  user: UserState
 }
 
 export type MultiModelMessageStyle = 'horizontal' | 'vertical' | 'fold' | 'grid'
@@ -131,7 +138,12 @@ const initialState: SettingsState = {
   notionPageNameKey: 'Name',
   thoughtAutoCollapse: true,
   notionAutoSplit: false,
-  notionSplitSize: 90
+  notionSplitSize: 90,
+  user: {
+    isLoggedIn: false,
+    username: '',
+    token: null
+  }
 }
 
 const settingsSlice = createSlice({
@@ -303,6 +315,9 @@ const settingsSlice = createSlice({
     },
     setNotionSplitSize: (state, action: PayloadAction<number>) => {
       state.notionSplitSize = action.payload
+    },
+    setUserState: (state, action: PayloadAction<Partial<UserState>>) => {
+      state.user = { ...state.user, ...action.payload }
     }
   }
 })
@@ -360,7 +375,8 @@ export const {
   setNotionPageNameKey,
   setThoughtAutoCollapse,
   setNotionAutoSplit,
-  setNotionSplitSize
+  setNotionSplitSize,
+  setUserState
 } = settingsSlice.actions
 
 export default settingsSlice.reducer
